pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["tradalhariri <trad.alhariri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.0"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
pytz = "^2021.3"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
